Fix MEDIUM findings from the full-repo audit + a deploy reliability gap
Security/routing:
- Caddy routed /order*, /favorite*, /admin* straight to the backend
  ahead of the SPA, so a hard refresh/bookmark of those pages hit a
  raw backend error instead of the app. Added an @spa_navigation
  matcher (GET + Accept: text/html) so real browser navigations reach
  the SPA shell while the frontend's own JSON API calls to the same
  paths still reach the backend - verified with a real caddy container
  against stand-in servers for all 4 cases (html GET, json GET, POST,
  catch-all).
- Session JWT was printed to the browser console on register/login/
  admin item creation (apiServise.js) - removed.
- deleteItemByName() passed the item's image URL to Cloudinary's
  destroy() instead of its public_id, so the image was never actually
  deleted - fixed to match deleteItemById()'s already-correct pattern.
- Deprecated jjwt 0.9-era API (Jwts.parser()/setSigningKey/signWith
  (Algorithm,String)) migrated to the modern parserBuilder()/Keys/
  signWith(Key) API in both services - same base64 decoding the old
  API already assumed, so this is a like-for-like behavior match, not
  a secret-format change. Was blocking any future jjwt version bump
  (those methods are removed entirely in 0.12.x).

Correctness:
- getAllOrderByEmail had no ORDER BY, but the frontend assumes the
  last row is the customer's current/open order - added ORDER BY id.
- Order.totalPrice was read via getDouble()+BigDecimal.valueOf()
  instead of getBigDecimal(), a float round-trip risk on money that
  feeds the WhatsApp order-total message.
- items-service's deleteItem(name) had its existence check commented
  out, so deleting a nonexistent name still reported success.
- Admin: negative prices passed validation (truthy-only check); the
  file-type guard was dead code from `!file instanceof File` operator
  precedence; the Update/Cancel button compared a function reference
  instead of the actual state var and never relabeled.
- SearchPage's fetch effect depended on `useParams()` itself (a new
  object every render) instead of `id`, causing an infinite refetch
  loop; its "not found" branch was dead code (checked truthiness of a
  value that started as [], always true) and the captured error was
  never shown.
- Favorite-heart state could be wrong and never self-correct for a
  logged-in user landing directly on /chocolates, /cakes, or /cookies
  (the fetch effect closed over a still-null currentUser from before
  the async "who am I" check resolved). Added a second effect keyed on
  [currentUser, isRequstToGetCurrentUserDone], matching the pattern
  Favorite.jsx already used correctly.

UX safety:
- Admin delete (item and user) had no confirmation step - added.

Efficiency:
- allOrderItemsInfo made 5 separate Feign calls per order item
  (name/image/description/veg/price) instead of 1 - fetch once, reuse.
  Also skips a null result instead of NPEing when items-service is
  briefly unavailable (fallback returns null).
- addToOrder queried getAllOrderOpen twice and reloaded the customer's
  entire order history (1 query + Feign fan-out per historical order)
  just to recompute the total of the one order that changed - now
  recomputes only that order's own items.
- Removed the Redis caching layer (CacheRepository/JedisConfig/
  RedisConfig, jedis + spring-boot-starter-data-redis deps, the
  spring.redis config block): fully implemented but never wired into
  any request path, and not provisioned in production (no redis
  container, SPRING_REDIS_IS_REDIS_SERVER_ACTIVE was already "false").
  Dead infrastructure removed rather than wiring in a new stateful
  service mid security-pass with no way to test cache-invalidation
  behavior end-to-end here.

Not attempted in this pass: the fragile "check the response string for
a substring" success/failure signal the audit flagged (10+ call sites
across both services, already caused one documented production bug
per an existing code comment). Fixing every site is too broad a blast
radius to do blind in one pass without integration-testing each
endpoint - flagging as a dedicated follow-up rather than a partial fix.

Deploy pipeline reliability (discovered mid-session, not an audit
finding): the HIGH-severity push's deploy job failed because the EC2
instance auto-shut-down mid-session; separately, once redeployed
manually, the new security headers didn't take effect - `docker
compose up -d` only restarts containers whose compose SERVICE
DEFINITION changed, and proxy's never does, so a Caddyfile-only edit
was silently never picked up (Caddy reads its config once at process
start, no hot-reload). Added an unconditional `restart proxy` step to
deploy.yml so this can't silently fail again.

Verified: both services compile clean. Caddyfile validated with the
real caddy:2-alpine binary, including a live functional test against
stand-in HTTP servers for the SPA-vs-API routing split. Merged
`docker compose config` confirms the new setup. The pre-existing,
unrelated Feign test-context failure (noted in the previous two
commits) is still present and still not a regression - flagging it
as worth a dedicated follow-up since it blocks `mvn test` entirely
right now, not chasing it further in this pass.

+++ security/src/main/java/com/example/security/repository/mapper/OrderMapper.java
@@ -17,7 +17,10 @@ public class OrderMapper implements RowMapper<Order> {
         order.setAddressShipping(rs.getString("address_shipping"));
         order.setStatus(Status.valueOf(rs.getString("order_status")));
         order.setOrderDate(rs.getDate("order_date").toLocalDate());
-        order.setTotalPrice(BigDecimal.valueOf(rs.getDouble("total_price")));
+        // Was rs.getDouble()+BigDecimal.valueOf() - a binary float round-trip
+        // on a DECIMAL(10,2) money column, unlike OrderItemsMapper/ItemMapper
+        // which correctly read the same column type via getBigDecimal().
+        order.setTotalPrice(rs.getBigDecimal("total_price"));
         return order;
     }
 }
